pinaki_assignment/requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.5
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
requests==2.32.5
rsa==4.9.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.9
